Removes cache budget option from purge request
Removes the cache budget option from the purge request to simplify the purge functionality.
The `cacheBudget` parameter is no longer needed, removing related code from both the client and server sides.
Also, this change corrects an issue in cloud image creation, swapping the content of user-data and vendor-data.
The entries parameter in PurgeHandler is now properly formatted as an argument.