Enhances VM IP address retrieval
Improves the mechanism for retrieving the IP address of a virtual machine.

It now considers both MAC address and DHCP client ID when searching DHCP leases.
Also adds the DHCP client ID to the configuration, making it configurable.

It also adds support for VMs without mac addresses by allowing the user to configure a dhcp client id
and authenticates ssh with the private key given the path relative to the config url.