Adds dependency build and packing scripts
Introduces scripts to build dependencies for macOS.

This includes downloading, patching, configuring, building, and installing the dependencies. It also adds a script to pack built libraries and headers into a library directory for easier consumption.
The build script builds the dependencies for a specified architecture and SDK.
It generates a meson cross file for cross-compilation and prevents host pkg-config contamination.
The packing script takes several architecture builds and combines them into single libraries.