Adds queue parameter for VM building
Introduces an optional `queue` parameter to the VM building process.
This allows for more control over the execution context of the IPSW installation,
enabling background processing and preventing UI blocking.
The `VirtualMachineWizard` is updated to leverage this new parameter, ensuring a
smoother and more responsive user experience during VM creation.