Converts qcow2 images to raw format
This change introduces a new qcow2 converter that converts qcow2 images to raw disk images.
It supports progress reporting via a callback function.
It removes reliance on `nativeimgutil` package.

The converter handles raw to raw conversions, qcow2 to raw conversions, and image resizing. It also creates a temporary file during conversion to handle cases where the source and destination are the same.