Adds agent installation feature
Enables the user to install an agent on a virtual machine from the UI.

This involves copying the necessary scripts to the VM and executing them. The agent allows to run services in background and perform privileged operations without the need of user password.

Also renames the icon to menubar.png and make other corrections in naming, copy and references.