Adds VNC server support
Adds the ability to run a VNC server for remote access to the virtual machine.

This change introduces a new "vnc" display mode, enabling users to connect to the VM via a VNC client.
The VNC server is started when the display mode is set to "vnc", either via command line or when the VM is launched from a service.

It also fixes an issue where the exclude paths for `cakedlib` were incorrect.