Adds getScreenSize functionality
Implements the ability to retrieve the screen size of the virtual machine.

This feature is exposed via gRPC and XPC to allow clients to query the current screen dimensions.
The VirtualMachine class now includes a getScreenSize method that returns the display size in pixels when running on macOS 14.0 or later.

The gRPC service and client are updated to include the getScreenSize method and associated message definitions.
The XPC service and client also now support getting the screen size.